1:11pm 3/25 [FIXED AND UPDATED]
A friend helped with the camera script. I feel as if it should be a time trial race, so the players are competing for the best time. I;m going to figure out a GUI. I will fix up the level for a better design and feel.  Let me know what you think.